+++ src/BootstrapTable.js
@@ -97,6 +97,7 @@ class BootstrapTable extends Component {
       keyField: keyField,
       colInfos: this.colInfos,
       multiColumnSearch: props.multiColumnSearch,
+      strictSearch: props.strictSearch,
       multiColumnSort: props.multiColumnSort,
       remote: this.props.remote
     });
@@ -108,9 +109,12 @@ class BootstrapTable extends Component {
     const sortName = options.defaultSortName || options.sortName;
     const sortOrder = options.defaultSortOrder || options.sortOrder;
     const searchText = options.defaultSearch;
+
     if (sortName && sortOrder) {
       this.store.setSortInfo(sortOrder, sortName);
-      this.store.sort();
+      if (!this.allowRemote(Const.REMOTE_SORT)) {
+        this.store.sort();
+      }
     }
 
     if (searchText) {
@@ -177,10 +181,11 @@ class BootstrapTable extends Component {
   }
 
   reset() {
+    const { pageStartIndex } = this.props.options;
     this.store.clean();
     this.setState({
       data: this.getTableData(),
-      currPage: 1,
+      currPage: Util.getFirstPage(pageStartIndex),
       expanding: [],
       sizePerPage: Const.SIZE_PER_PAGE_LIST[0],
       selectedRowKeys: this.store.getSelectedRowKeys(),
@@ -353,7 +358,7 @@ class BootstrapTable extends Component {
         { toolBar }
         { showPaginationOnTop ? pagination : null }
         <div ref='table'
-            className={ classSet('react-bs-table', this.props.tableContainerClass) }
+            className={ classSet('react-bs-table', { 'react-bs-table-bordered': this.props.bordered }, this.props.tableContainerClass) }
             style={ { ...style, ...this.props.tableStyle } }
             onMouseEnter={ this.handleMouseEnter }
             onMouseLeave={ this.handleMouseLeave }>
@@ -410,6 +415,7 @@ class BootstrapTable extends Component {
             withoutNoDataText={ this.props.options.withoutNoDataText }
             expanding={ this.state.expanding }
             onExpand={ this.handleExpandRow }
+            onlyOneExpanding={ this.props.options.onlyOneExpanding }
             beforeShowError={ this.props.options.beforeShowError }
             keyBoardNav={ this.props.keyBoardNav }
             onNavigateCell={ this.handleNavigateCell }
@@ -498,18 +504,7 @@ class BootstrapTable extends Component {
       return;
     }
 
-    // We calculate an offset here in order to properly fetch the indexed data,
-    // despite the page start index not always being 1
-    let normalizedPage;
-    if (pageStartIndex !== undefined) {
-      const offset = Math.abs(Const.PAGE_START_INDEX - pageStartIndex);
-      normalizedPage = page + offset;
-    } else {
-      normalizedPage = page;
-    }
-
-    const result = this.store.page(normalizedPage, sizePerPage).get();
-
+    const result = this.store.page(Util.getNormalizedPage(pageStartIndex, page), sizePerPage).get();
     this.setState({ data: result, reset: false });
   }
 
@@ -542,8 +537,6 @@ class BootstrapTable extends Component {
     let { x, y, currPage } = this.state;
     x += offSetX;
     y += offSetY;
-    // currPage += 1;
-    // console.log(currPage);
 
     const columns = this.store.getColInfos();
     const visibleRowSize = this.state.data.length;
@@ -666,16 +659,17 @@ class BootstrapTable extends Component {
 
   handleShowOnlySelected = () => {
     this.store.ignoreNonSelected();
+    const { pageStartIndex } = this.props.options;
     let result;
     if (this.props.pagination) {
-      result = this.store.page(1, this.state.sizePerPage).get();
+      result = this.store.page(Util.getNormalizedPage(pageStartIndex), this.state.sizePerPage).get();
     } else {
       result = this.store.get();
     }
     this.setState({
       data: result,
       reset: false,
-      currPage: this.props.options.pageStartIndex || Const.PAGE_START_INDEX
+      currPage: Util.getFirstPage(pageStartIndex)
     });
   }
 
@@ -858,14 +852,14 @@ class BootstrapTable extends Component {
   }
 
   handleFilterData = filterObj => {
-    const { onFilterChange } = this.props.options;
+    const { onFilterChange, pageStartIndex } = this.props.options;
     if (onFilterChange) {
       const colInfos = this.store.getColInfos();
       onFilterChange(filterObj, colInfos);
     }
 
     this.setState({
-      currPage: this.props.options.pageStartIndex || Const.PAGE_START_INDEX,
+      currPage: Util.getFirstPage(pageStartIndex),
       reset: false
     });
 
@@ -888,7 +882,7 @@ class BootstrapTable extends Component {
 
     if (this.props.pagination) {
       const { sizePerPage } = this.state;
-      result = this.store.page(1, sizePerPage).get();
+      result = this.store.page(Util.getNormalizedPage(pageStartIndex), sizePerPage).get();
     } else {
       result = this.store.get();
     }
@@ -943,14 +937,14 @@ class BootstrapTable extends Component {
     if (this.refs.toolbar) {
       this.refs.toolbar.setSearchInput(searchText);
     }
-    const { onSearchChange } = this.props.options;
+    const { onSearchChange, pageStartIndex } = this.props.options;
     if (onSearchChange) {
       const colInfos = this.store.getColInfos();
       onSearchChange(searchText, colInfos, this.props.multiColumnSearch);
     }
 
     this.setState({
-      currPage: this.props.options.pageStartIndex || Const.PAGE_START_INDEX,
+      currPage: Util.getFirstPage(pageStartIndex),
       reset: false
     });
 
@@ -973,7 +967,7 @@ class BootstrapTable extends Component {
     let result;
     if (this.props.pagination) {
       const { sizePerPage } = this.state;
-      result = this.store.page(1, sizePerPage).get();
+      result = this.store.page(Util.getNormalizedPage(pageStartIndex), sizePerPage).get();
     } else {
       result = this.store.get();
     }
@@ -1057,6 +1051,7 @@ class BootstrapTable extends Component {
             field: props.dataField,
             hiddenOnInsert: props.hiddenOnInsert,
             keyValidator: props.keyValidator,
+            customInsertEditor: props.customInsertEditor,
             // when you want same auto generate value and not allow edit, example ID field
             autoValue: props.autoValue || false,
             // for create editor, no params for column.editable() indicate that editor for new row
@@ -1071,6 +1066,7 @@ class BootstrapTable extends Component {
           name: children.props.headerText || children.props.children,
           field: children.props.dataField,
           editable: children.props.editable,
+          customInsertEditor: children.props.customInsertEditor,
           hiddenOnInsert: children.props.hiddenOnInsert,
           keyValidator: children.props.keyValidator
         } ];
@@ -1268,11 +1264,11 @@ class BootstrapTable extends Component {
       const { sizePerPage } = this.state;
 
       if (atTheBeginning) {
-        const firstPage = this.props.options.pageStartIndex || Const.PAGE_START_INDEX;
-        result = this.store.page(firstPage, sizePerPage).get();
+        const { pageStartIndex } = this.props.options;
+        result = this.store.page(Util.getNormalizedPage(pageStartIndex), sizePerPage).get();
         this.setState({
           data: result,
-          currPage: firstPage,
+          currPage: Util.getFirstPage(pageStartIndex),
           reset: false
         });
       } else {
@@ -1342,6 +1338,8 @@ BootstrapTable.propTypes = {
   insertRow: PropTypes.bool,
   deleteRow: PropTypes.bool,
   search: PropTypes.bool,
+  multiColumnSearch: PropTypes.bool,
+  strictSearch: PropTypes.bool,
   columnFilter: PropTypes.bool,
   trClassName: PropTypes.any,
   tableStyle: PropTypes.object,
@@ -1429,6 +1427,7 @@ BootstrapTable.propTypes = {
     expandRowBgColor: PropTypes.string,
     expandBy: PropTypes.string,
     expanding: PropTypes.array,
+    onlyOneExpanding: PropTypes.bool,
     beforeShowError: PropTypes.func,
     printToolBar: PropTypes.bool
   }),
@@ -1491,6 +1490,7 @@ BootstrapTable.defaultProps = {
   deleteRow: false,
   search: false,
   multiColumnSearch: false,
+  strictSearch: undefined,
   multiColumnSort: 1,
   columnFilter: false,
   trClassName: '',
@@ -1544,7 +1544,7 @@ BootstrapTable.defaultProps = {
     nextPageTitle: Const.NEXT_PAGE_TITLE,
     firstPageTitle: Const.FIRST_PAGE_TITLE,
     lastPageTitle: Const.LAST_PAGE_TITLE,
-    pageStartIndex: undefined,
+    pageStartIndex: 1,
     searchDelayTime: undefined,
     exportCSVText: Const.EXPORT_CSV_TEXT,
     insertText: Const.INSERT_BTN_TEXT,
@@ -1572,6 +1572,7 @@ BootstrapTable.defaultProps = {
     expandRowBgColor: undefined,
     expandBy: Const.EXPAND_BY_ROW,
     expanding: [],
+    onlyOneExpanding: false,
     beforeShowError: undefined,
     printToolBar: true
   },