+++ src/TableEditColumn.js
@@ -16,8 +16,12 @@ class TableEditColumn extends Component {
     };
   }
 
+  valueShortCircuit(value) {
+    return value === null || typeof value === 'undefined' ? '' : value;
+  }
+
   handleKeyPress = e => {
-    if (e.keyCode === 13) {
+    if (e.keyCode === 13 || e.keyCode === 9) {
       // Pressed ENTER
       const value = e.currentTarget.type === 'checkbox' ?
                       this._getCheckBoxValue(e) : e.currentTarget.value;
@@ -25,13 +29,15 @@ class TableEditColumn extends Component {
       if (!this.validator(value)) {
         return;
       }
-      this.props.completeEdit(value, this.props.rowIndex, this.props.colIndex);
+      if (e.keyCode === 13) {
+        this.props.completeEdit(value, this.props.rowIndex, this.props.colIndex);
+      } else {
+        this.props.onTab(this.props.rowIndex + 1, this.props.colIndex + 1, 'tab', e);
+        e.preventDefault();
+      }
     } else if (e.keyCode === 27) {
       this.props.completeEdit(
         null, this.props.rowIndex, this.props.colIndex);
-    } else if (e.keyCode === 9) {
-      this.props.onTab(this.props.rowIndex + 1, this.props.colIndex + 1, 'tab', e);
-      e.preventDefault();
     } else if (e.type === 'click' && !this.props.blurToSave) {  // textarea click save button
       const value = e.target.parentElement.firstChild.value;
       if (!this.validator(value)) {
@@ -48,7 +54,7 @@ class TableEditColumn extends Component {
       const value = e.currentTarget.type === 'checkbox' ?
                       this._getCheckBoxValue(e) : e.currentTarget.value;
       if (!this.validator(value)) {
-        return;
+        return false;
       }
       this.props.completeEdit(
           value, this.props.rowIndex, this.props.colIndex);
@@ -168,18 +174,18 @@ class TableEditColumn extends Component {
     editable.placeholder && (attr.placeholder = editable.placeholder);
 
     const editorClass = classSet({ 'animated': shakeEditor, 'shake': shakeEditor });
+    fieldValue = fieldValue === 0 ? '0' : fieldValue;
     let cellEditor;
     if (customEditor) {
       const customEditorProps = {
         row,
         ...attr,
-        defaultValue: fieldValue || '',
+        defaultValue: this.valueShortCircuit(fieldValue),
         ...customEditor.customEditorParameters
       };
       cellEditor = customEditor.getElement(this.handleCustomUpdate, customEditorProps);
     } else {
-      fieldValue = fieldValue === 0 ? '0' : fieldValue;
-      cellEditor = editor(editable, attr, format, editorClass, fieldValue || '');
+      cellEditor = editor(editable, attr, format, editorClass, this.valueShortCircuit(fieldValue));
     }
 
     if (isFocus) {